pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fmri2image"
version = "0.1.0"
description = "fMRI→Image reconstruction pipeline (Hydra, DVC, Lightning, W&B)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.3",
  "pytorch-lightning>=2.4",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "dvc[s3]>=3.50",
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.5",
  "einops>=0.8",
  "tqdm>=4.66",
  "wandb>=0.17",
  "rich>=13.7",
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "mypy>=1.11",
  "ruff>=0.6",
  "xarray>=2024.6",
  "nibabel>=5.2",
  "nilearn>=0.10",
  "h5py>=3.11",
]

[tool.ruff]
line-length = 100
select = ["E","F","I","B","UP"]
exclude = ["data","build",".venv"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = false

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
